backend/requirements.txt
fastapi==0.122.0
uvicorn== 0.38.0
celery==5.5.3
redis==7.1.0
pillow==12.0.0
transformers==4.57.3
torch==2.9.1    
python-multipart==0.0.20
gdown==5.2.0